+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## TypeScriptESLintPlugin v1.0.5
+  - Allow the `typescript-eslint-plugin` to run as a global plugin
+  - Provide the functionality to disable the logging by setting `logLevel` to `none`
+
+[Show differences](https://github.com/manuth/TypeScriptESLintPlugin/compare/v1.0.4..v1.0.5)
+
 ## TypeScriptESLintPlugin v1.0.4
   - Fix incorrect messages
   - Add missing dependencies